+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 - Add support for the rewatch build system for incremental compilation. https://github.com/rescript-lang/rescript-vscode/pull/965
 - Add support for Linux ARM64
 - Statically linked Linux binaries
+- Emit `%todo` instead of `failwith("TODO")` when we can (ReScript >= v11.1). https://github.com/rescript-lang/rescript-vscode/pull/981
+- Complete `%todo`. https://github.com/rescript-lang/rescript-vscode/pull/981
 
 ## 1.50.0